:white_check_mark: add require tests
